Studio: Moved MultiWellplate into org.micromanager.data
