+++ mmstudio/src/main/java/org/micromanager/data/MultiWellPlate.java
@@ -1,6 +1,7 @@
-package org.micromanager;
+package org.micromanager.data;
 
 
+import org.micromanager.PropertyMap;
 
 /**
  * Metadata describing a multi-well plate
